Avoid rebuilding database fixtures for every test run.
This reduces the runtime of database-bound tests by about 40%
(on my system, from 4:55 to 2:47).

+++ tests/phpunit/MediaWikiTestCase.php
@@ -134,13 +134,18 @@ abstract class MediaWikiTestCase extends PHPUnit_Framework_TestCase {
 			if ( !self::$dbSetup ) {
 				// switch to a temporary clone of the database
 				self::setupTestDB( $this->db, $this->dbPrefix() );
+				$this->addCoreDBData();
 
 				if ( ( $this->db->getType() == 'oracle' || !self::$useTemporaryTables ) && self::$reuseDB ) {
 					$this->resetDB();
 				}
 			}
-			$this->addCoreDBData();
-			$this->addDBData();
+
+			if ( $this->oncePerClass() ) {
+				$this->addDBData();
+			}
+
+			$this->prepareDB();
 			$needsResetDB = true;
 		}
 
@@ -149,6 +154,22 @@ abstract class MediaWikiTestCase extends PHPUnit_Framework_TestCase {
 		if ( $needsResetDB ) {
 			$this->resetDB();
 		}
+	}
+
+	/**
+	 * @return boolean
+	 */
+	private function oncePerClass() {
+		// Remember current test class in the database connection,
+		//so we know when we need to run addData.
+
+		$class = get_class( $this );
+
+		$first = !isset( $this->db->_hasDataForTestClass )
+			|| $this->db->_hasDataForTestClass !== $class;
+
+		$this->db->_hasDataForTestClass = $class;
+		return $first;
 	}
 
 	/**
@@ -527,10 +548,18 @@ abstract class MediaWikiTestCase extends PHPUnit_Framework_TestCase {
 	}
 
 	/**
-	 * Stub. If a test needs to add additional data to the database, it should
-	 * implement this method and do so
+	 * Stub. If a test suite needs to add additional data to the database, it should
+	 * implement this method and do so. This method is called once per test suite.
 	 *
-	 * @since 1.18
+	 * Note data added by this method may be removed by resetDB() depending on
+	 * the contents of $tablesUsed.
+	 *
+	 * To control the database reset between tests, override prepareDB() or resetDB().
+	 *
+	 * @see prepareDB()
+	 * @see resetDB()
+	 *
+	 * @since 1.27 (invocation contract changed)
 	 */
 	public function addDBData() {
 	}
@@ -671,27 +700,39 @@ abstract class MediaWikiTestCase extends PHPUnit_Framework_TestCase {
 	}
 
 	/**
-	 * Empty all tables so they can be repopulated for tests
+	 * Stub. Subclasses may override this to prepare the database.
+	 * Called before every test run (test function or data set).
+	 *
+	 * @see addDBData()
+	 * @see resetDB()
+	 *
+	 * @since 1.27
 	 */
-	private function resetDB() {
+	protected function prepareDB() {
+	}
+
+	/**
+	 * Empty all tables so they can be repopulated for tests.
+	 * Subclasses may override this to perform additional cleanup.
+	 * Called after every test run (test function or data set).
+	 *
+	 * @see addDBData()
+	 * @see prepareDB()
+	 *
+	 * @since 1.27 (became available as protected method)
+	 */
+	protected function resetDB() {
 		if ( $this->db ) {
-			if ( $this->db->getType() == 'oracle' ) {
-				if ( self::$useTemporaryTables ) {
-					wfGetLB()->closeAll();
-					$this->db = wfGetDB( DB_MASTER );
-				} else {
-					foreach ( $this->tablesUsed as $tbl ) {
-						if ( $tbl == 'interwiki' ) {
-							continue;
-						}
-						$this->db->query( 'TRUNCATE TABLE ' . $this->db->tableName( $tbl ), __METHOD__ );
-					}
+			$truncate = in_array( $this->db->getType(), [ 'oracle', 'mysql' ] );
+			foreach ( $this->tablesUsed as $tbl ) {
+				// TODO: reset interwiki and user tables to their original content.
+				if ( $tbl == 'interwiki' || $tbl == 'user' ) {
+					continue;
 				}
-			} else {
-				foreach ( $this->tablesUsed as $tbl ) {
-					if ( $tbl == 'interwiki' || $tbl == 'user' ) {
-						continue;
-					}
+
+				if ( $truncate ) {
+					$this->db->query( 'TRUNCATE TABLE ' . $this->db->tableName( $tbl ), __METHOD__ );
+				} else {
 					$this->db->delete( $tbl, '*', __METHOD__ );
 				}
 			}
@@ -1283,4 +1324,5 @@ abstract class MediaWikiTestCase extends PHPUnit_Framework_TestCase {
 	public static function wfResetOutputBuffersBarrier( $buffer ) {
 		return $buffer;
 	}
+
 }
